provisioner: Use tftp.service on SLE12 instead of xinetd

+++ chef/cookbooks/provisioner/recipes/setup_base_images.rb
@@ -279,33 +279,59 @@ cookbook_file "/etc/tftpd.conf" do
 end
 
 if node[:platform_family] == "suse"
-  service "tftp" do
-    # just enable, don't start (xinetd takes care of it)
-    enabled node[:provisioner][:enable_pxe] ? true : false
-    action node[:provisioner][:enable_pxe] ? "enable" : "disable"
-  end
+  if node[:platform] == "suse" && node[:platform_version].to_f < 12.0
+    service "tftp" do
+      # just enable, don't start (xinetd takes care of it)
+      enabled node[:provisioner][:enable_pxe] ? true : false
+      action node[:provisioner][:enable_pxe] ? "enable" : "disable"
+    end
 
-  # NOTE(toabctl): stop for tftp does not really help. the process gets started
-  # by xinetd and has a default timeout of 900 seconds which triggers when no
-  # new connections start in this period. So kill the process here
-  execute "kill in.tftpd process" do
-    command "pkill in.tftpd"
-    not_if { node[:provisioner][:enable_pxe] }
-    returns [0, 1]
-  end
+    # NOTE(toabctl): stop for tftp does not really help. the process gets started
+    # by xinetd and has a default timeout of 900 seconds which triggers when no
+    # new connections start in this period. So kill the process here
+    execute "kill in.tftpd process" do
+      command "pkill in.tftpd"
+      not_if { node[:provisioner][:enable_pxe] }
+      returns [0, 1]
+    end
 
-  service "xinetd" do
-    running node[:provisioner][:enable_pxe] ? true : false
-    enabled node[:provisioner][:enable_pxe] ? true : false
-    action node[:provisioner][:enable_pxe] ? ["enable", "start"] : ["disable", "stop"]
-    supports reload: true
-    subscribes :reload, resources(service: "tftp"), :immediately
-  end
+    service "xinetd" do
+      running node[:provisioner][:enable_pxe] ? true : false
+      enabled node[:provisioner][:enable_pxe] ? true : false
+      action node[:provisioner][:enable_pxe] ? ["enable", "start"] : ["disable", "stop"]
+      supports reload: true
+      subscribes :reload, resources(service: "tftp"), :immediately
+    end
 
-  template "/etc/xinetd.d/tftp" do
-    source "tftp.erb"
-    variables( tftproot: tftproot )
-    notifies :reload, resources(service: "xinetd")
+    template "/etc/xinetd.d/tftp" do
+      source "tftp.erb"
+      variables(tftproot: tftproot)
+      notifies :reload, resources(service: "xinetd")
+    end
+  else
+    template "/etc/systemd/system/tftp.service" do
+      source "tftp.service.erb"
+      owner "root"
+      group "root"
+      mode "0644"
+      variables(tftproot: tftproot, admin_ip: admin_ip)
+    end
+
+    service "tftp.service" do
+      if node[:provisioner][:enable_pxe]
+        action ["enable", "start"]
+        subscribes :restart, resources("cookbook_file[/etc/tftpd.conf]")
+        subscribes :restart, resources("template[/etc/systemd/system/tftp.service]")
+      else
+        action ["disable", "stop"]
+      end
+    end
+
+    bash "reload systemd after tftp.service update" do
+      code "systemctl daemon-reload"
+      action :nothing
+      subscribes :run, resources(template: "/etc/systemd/system/tftp.service"), :immediately
+    end
   end
 else
   template "/etc/bluepill/tftpd.pill" do
